Update README.md
This repository contains the code for a responsive Food Delivery Website built with HTML, CSS, Bootstrap, and JavaScript. The website is designed to provide an intuitive and visually appealing interface for users to browse and order food from various restaurants. It is optimized for viewing on multiple devices, ensuring a seamless experience across desktops, tablets, and mobile phones.